Backend/requirements.txt
Flask
python-dotenv
Flask-restful
flask_sqlalchemy
flask-jwt-extended
flask-mail